% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{Function_2}
\alias{Function_2}
\title{Function 2 creates plots based on the class input in the function}
\usage{
Function_2(dataset, class_name)
}
\arguments{
\item{dataset}{is a data frame of the users choosing}

\item{class_name}{}
}
\value{
returns a histogram of the variables that are of class class_name if the class is numeric, double, or integer, returns a bar plot of the variables that are of class class_name if the class is factor, logical, or character
}
\description{
Function 2 creates plots based on the class input in the function
}
\examples{
A <- c("J","e","d","H","a","m","m","o","u","d")
B <- c(1,2,3,4,1,2,3,4,4,6)
C <- c(1,2,3,2,1,2,1,2,3,3)
D <- c("a","b","c","c","d","d","e","f","g","f")
E <- c(1,4,5,6,7,4,5,42,8,11)
f <-c(TRUE,TRUE,FALSE,FALSE,FALSE,TRUE,TRUE,FALSE,FALSE,TRUE)
Example1 <- data.frame(A,B,C,D,E,f)
A1 <- c("g","f","d","d","l","n","z","w","ud","d")
B2 <- c(11,22,33,44,11,22,33,44,44,66)
C3 <- c(1,4,9,4,1,4,1,8,33,1089)
D4 <- c("a","b","c","c","d","d","e","f","g","f")
E5 <- c(FALSE,FALSE,FALSE,TRUE,TRUE,TRUE,FALSE,FALSE,TRUE,TRUE)
f6 <-c(TRUE,TRUE,FALSE,TRUE,FALSE,TRUE,FALSE,FALSE,FALSE,FALSE)
Example2 <- data.frame(A1,B2,C3,D4,E5,f6)
examp1A <- Function_2(Example1, "numeric")
examp1A
examp2A <- Function_2(Example2, "logical")
examp2A
}
