% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW04Jh5764a-package.R
\docType{package}
\name{HW04Jh5764a-package}
\alias{HW04Jh5764a}
\alias{HW04Jh5764a-package}
\title{HW04Jh5764a: Dataframe Manipulation and Visualization}
\description{
This package contains two functions. One where you can enter a data frame and get different results and one where you can enter a data frame and get differnet plots
}
\keyword{internal}
